Add The Hack (#1)
This was supposed to be a much smaller PR with just the contents of
TheHack and its tests. However after starting to test generics it became
clear that we needed the `resolveType` logic from Mockspresso2, so we
import that as well.

We make one modification to that logic: Mockspresso2 is only concerned
with class constructors, so it did not include support for generic
methods. We tried to add that here but we can only support methods where
the generics only have a single upper bound.

This does NOT affect generic classes, those should be fine. This is only
a problem for methods with their own generics (i.e. type variables that
are not part of the class)